[REFACT] select를 이용해서 서버 데이터를 가공하여 반환하기 #NADA-52
